Code Reorg: Split vex package
This commit rearranges the code into new _functions.go files to
get them out of the main vex.go file. Tests are moved into their
corresponding files but no other changes are done.
